requirements.txt
djangorestframework~=3.13.1
django==3.2
psycopg2-binary
requests~=2.23.0
django-crontab
pandas~=1.4.3
SQLAlchemy~=1.3.18
django-cors-headers
sentry-sdk
pytz~=2022.1
python-decouple
python-dateutil==2.8.1


# Deprecated
django-composite-foreignkey==1.1.0
numpy~=1.23.1